+++ CHANGELOG.md
@@ -10,11 +10,12 @@
 - new `queryParser` config option to allow custom query string parsing
 - better errors
 - automatically lowercase header names
+- `PATCH` method
 
 ## Fixes
 
 - absolute paths as `baseUrl` is now supported in fetch backend
-- don't re-use same requests based upon URL to allow for concurrent POST requests
+- don't re-use same requests based upon URL to allow for concurrent `POST` requests
 
 # 0.5.0